move sfen reader to separate file
